pyproject.toml
[tool.poetry]
name = "pii-extraction-api"
version = "0.1.0"
description = ""
authors = ["Evgeniy Bliacher <evgeniy.bliacher@alterahealth.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.9.2"
azure-ai-textanalytics = "^5.3.0"
pymonad = "^2.4.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.31.0"}
assertpy = "^1.1"
asyncio = "^3.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
